src/api/requirements.txt
boto3
git+https://gitlab.com/obuilds/public/pytube@ob-v1
Flask
flask-cors
flake8
python-dotenv
moviepy